fix: wrong method using in forgot password method, find user role then delete instead of creating new instance, all endpoints in user controller return response dto add: get user role method

+++ Services/Implementations/UserService.cs
@@ -92,13 +92,11 @@ public class UserService : IUserService
         {
             throw new NotFoundException("Role not found");
         }
-        UserRole userRole = new UserRole
+        UserRole? userRole = await _userRepository.GetUserRoleAsync(userId, roleId);
+        if (userRole == null)
         {
-            User = user,
-            Role = role,
-            UserId = userId,
-            RoleId = roleId
-        };
+            throw new NotFoundException("User role not found");
+        }
         await _userRepository.RemoveRoleFromUserAsync(userRole);
     }
 }
